add Quran UI module with loading and fullscreen functionality
